Made path references more generic
Removed some of the fully-qualified paths to files in order to allow for
the manifests to be processed by a Puppet Windows agent - at least when
some Unix compatibility layer like cygwin has been loaded on the agent
system.

Added external_fact_hostname.bat variant of the external_fact_hostname
fact for Windows.